rename: .js > .ts 파일로 변경
