+++ src/component/TodoCount.ts
@@ -1,4 +1,4 @@
-import validation from "../utils/validation.js";
+import validation from "../util/validation.js";
 
 export default function TodoCount({ $target, initialCount }) {
   validation.newTarget(new.target);
