+++ src/component/TodoForm.ts
@@ -1,4 +1,4 @@
-import validation from "../utils/validation.js";
+import validation from "../util/validation.js";
 
 export default function TodoForm({ $target, onSubmit }) {
   validation.newTarget(new.target);
